requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==2.1.0
Django==4.2.5
django-crispy-forms==2.0
django-heroku==0.3.1
django-phonenumber-field==7.2.0
django-taggit==4.0.0
django-tinymce==3.6.1
gunicorn==21.2.0
Markdown==3.4.4
packaging==23.2
phonenumbers==8.13.22
Pillow==10.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
